Python projects/Check EMAG for Discounts - Flask Web App/requirements.txt
aniso8601==9.0.1
bcrypt==3.2.0
beautifulsoup4==4.9.3
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.1
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.2.1
Flask-Login==0.5.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==0.15.1
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
jwt==1.2.0
MarkupSafe==2.0.1
mongoengine==0.23.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.4
python-dotenv==0.18.0
pytz==2021.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
Werkzeug==2.0.1
WTForms==2.3.3
